config: move firewall-filter warn validators to own file
Pure code-motion for #4845. Move the firewall/filter WARN validators
(validateFilterLossPriorityWarnings, validateFirewallInterfaceSpecificWarnings,
validateLo0FilterKernelMirrorWarnings, validateFilterNoCatchAllWarnings) and
their term/catch-all helpers verbatim from compiler_validate_warn.go into
compiler_validate_warn_firewall.go.

Bodies are byte-identical; only the file each function lives in and the
per-file import block change. Same package config, same signatures. No
behavior change.
